# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts App",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "A RESTful API for managing contacts, including authentication and user management."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://nodejs-hw-mongodb-et1c.onrender.com"
    }
  ],
...
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "operationId": "refreshToken",
        "description": "Generates a new access token using a valid refresh token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Refresh token used to generate a new access token.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.xedzdWIiOiPnwsh0NTY3ODkwIiwibmFtZSI6IkpNSOugRG9lIiwiaALDUjoxNTE2MjM5MDIyfQ.FSFvvKQ66B8w87W5Y1Pfe0xgX1MP8",
                    "description": "A valid refresh token issued during authentication."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully generated a new access token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200,
                      "description": "HTTP status code indicating successful token refresh."
                    },
                    "message": {
                      "type": "string",
                      "example": "Access token refreshed successfully",
                      "description": "Confirmation message of successful token refresh."
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "accessToken"
                      ],
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                          "description": "New JWT access token for authenticated requests."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or missing refresh token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400,
                      "description": "HTTP status code indicating a bad request."
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid refresh token",
                      "description": "Description of the error."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access, invalid or expired refresh token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401,
                      "description": "HTTP status code indicating unauthorized access."
                    },
                    "message": {
                      "type": "string",
                      "example": "Refresh token expired or invalid",
                      "description": "Error message indicating token issue."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/auth/reset-pwd": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset user password",
        "operationId": "resetPassword",
        "description": "Allows a user to reset their password using a valid reset token.",
        "security": [],
        "requestBody": {
          "description": "Request to reset the password.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "resetToken",
                  "newPassword"
                ],
                "properties": {
                  "resetToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                    "description": "Token for password reset, received via email."
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "example": "Nikorest123!",
                    "description": "The new password for the user account."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully reset the user's password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200,
                      "description": "HTTP status code indicating password reset success."
                    },
                    "message": {
                      "type": "string",
                      "example": "Password reset successfully",
                      "description": "Confirmation message for successful password reset."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or missing reset token or password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400,
                      "description": "HTTP status code indicating a bad request."
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid reset token or password format",
                      "description": "Detailed error message."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access, expired or invalid reset token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401,
                      "description": "HTTP status code indicating unauthorized access."
                    },
                    "message": {
                      "type": "string",
                      "example": "Reset token expired or invalid",
                      "description": "Error message indicating token issue."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500,
                  "description": "HTTP status code indicating an internal server error."
                },
                "message": {
                  "type": "string",
                  "example": "Internal server error.",
                  "description": "A generic message indicating that the server encountered an unexpected error."
                },
                "errorCode": {
                  "type": "string",
                  "example": "SERVER_ERROR",
                  "description": "A specific error code to help identify the nature of the internal issue."
                },
                "method": {
                  "type": "string",
                  "example": "POST",
                  "description": "The HTTP method used when the error occurred."
                }
              }
            }
          }
        }
      },
...
    }
  }
}